Add name length check to the HipChatHandler
As explained in its docblock, the validation function will fall back to
a simple strlen check if mb_strlen is not available. However, that very
specific case is not one that was deemed important to support, as anyone
using UTF-8 in their code, without having mbstring available will
probably have bigger problems to worry about.

Fixes #289